Refactored js components into separate files
